Design fixes
