+++ src/assembly/mod.rs
@@ -1,3 +1,2 @@
 pub mod explication;
 pub mod parser;
-mod explication_context;
